@@ -1,5 +1,4 @@
 import { TestBed, async } from '@angular/core/testing';
-import { MaterialModule } from '@angular/material';
 import { RouterTestingModule } from '@angular/router/testing';
 
 import { AppComponent } from './app.component';
@@ -7,13 +6,8 @@ import { AppComponent } from './app.component';
 describe('AppComponent', () => {
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      imports: [
-        MaterialModule,
-        RouterTestingModule,
-      ],
-      declarations: [
-        AppComponent
-      ],
+      declarations: [ AppComponent ],
+      imports: [ RouterTestingModule ],
     }).compileComponents();
   }));
 
