@@ -5,5 +5,5 @@
 
 export const environment = {
   production: false,
-  server: 'http://localhost:8000'
+  server: 'http://localhost:8000',
 };
